% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_lookup.R
\name{div_lookup}
\alias{div_lookup}
\title{Lookup table for concentration/diversification variables}
\usage{
div_lookup(x, nms = c("_num", "_hhi", "_sd"))
}
\arguments{
\item{x}{tibble with the data}

\item{nms}{names used to select the concentration/diversification variables}
}
\value{
a tibble
}
\description{
Lookup table for concentration/diversification variables
}
