% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversifcation-package.R
\docType{package}
\name{diversification-package}
\alias{diversification}
\alias{diversification-package}
\title{diversification: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/diversification}
  \item Report bugs at \url{https://github.com/f-dallolio/diversification/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
