% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_div_fn.R
\name{make_div_fn}
\alias{make_div_fn}
\title{Create concentration/diversification measures}
\usage{
make_div_fn(.data = panel_df, .fn, .var, .id = 1)
}
\arguments{
\item{.data}{the data frame with the data.}

\item{.fn}{the function used. Must be one of \code{c("cfx", "nfx", "tau", "dfx", "dau")}.}

\item{.var}{the variable being transformed. Must be one of ...}

\item{.id}{...}
}
\value{
a numeric vector.
}
\description{
Create concentration/diversification measures
}
