% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tau.R
\name{make_tau}
\alias{make_tau}
\title{Diversifications Measures}
\usage{
make_tau(num, hhi)
}
\arguments{
\item{num}{a numeric vector.}

\item{hhi}{a numeric vector.}
}
\value{
a numeric vector.
}
\description{
Diversifications Measures
}
