% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_by_key.R
\name{mutate_by_key}
\alias{mutate_by_key}
\title{Mutate by key only for tsibble}
\usage{
mutate_by_key(
  .data,
  .and_by = NULL,
  ...,
  .keep = c("all", "used", "unused", "none"),
  .before = NULL,
  .after = NULL
)
}
\arguments{
\item{.after}{}
}
\description{
Mutate by key only for tsibble
}
