% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by_key.R
\name{summarise_by_key}
\alias{summarise_by_key}
\title{Summary by key for tsibble}
\usage{
summarise_by_key(.data, ..., .and_by = NULL, .groups = NULL)
}
\arguments{
\item{.groups}{}
}
\description{
Summary by key for tsibble
}
